Create Compiler_Version.java
You are converting an old code for a new version of the compiler.

In the old code we have used "->" for pointers. But now we have to replace each "->" with a ".". But this replacement shouldn't be done inside commments. A comment is a string that starts with "//" and terminates at the end of the line.